[WEB-78] ♻️ refactor: AppHeaderBottomBar 상태 변경
